requirements.txt
ccxt
discord-webhook
empyrical
influxdb-client
jupyter
matplotlib
matplotlib-inline
mypy-extensions
notebook
numba
numpy
pandas
pipe
plotly
psutil
pyarrow
python-dateutil
python-dotenv
PyYAML
ray==2.0.0
requests
scikit-learn
scipy
streamlit
vectorbt